requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
httpx>=0.26.0
alembic>=1.13.1
python-multipart>=0.0.6
openpyxl>=3.1.2
# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
